impl queue depending on pinned concurrent collection
