+++ Cargo.toml
@@ -8,6 +8,7 @@ publish = ["artifactory"]
 # orx-concurrent-bag = { version = "3.1.0", registry = "artifactory" }
 orx-concurrent-bag = { path = "../orx-concurrent-bag", registry = "artifactory" }
 orx-concurrent-iter = { path = "../orx-concurrent-iter", registry = "artifactory" }
+orx-pinned-concurrent-col = { path = "../orx-pinned-concurrent-col", registry = "artifactory" }
 orx-concurrent-vec = { version = "3.8.0", registry = "artifactory" }
 orx-pinned-vec = { version = "3.17.0", registry = "artifactory" }
 orx-split-vec = { version = "3.19.0", registry = "artifactory" }
