+++ src/tests/multiple_items.rs
@@ -1,4 +1,4 @@
-use crate::queue::Queue;
+use crate::queue_fixed::Queue;
 use orx_concurrent_bag::*;
 use std::fmt::Debug;
 use test_case::test_matrix;
